pyproject.toml
[tool.poetry]
name = "minecraft_analytics"
version = "0.1.0"
description = "Tools for Minecraft research"
authors = ["Chris Proctor <chrisp@buffalo.edu>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.25.1"
invoke = "^1.5.0"
pandas = "^1.2.5"
ics = "^0.7"
pyaml = "^20.4.0"
jupyter = "^1.0.0"
python-dateutil = "^2.8.2"
matplotlib = "^3.4.3"
seaborn = "^0.11.2"
scikit-learn = "^1.0.1"
statsmodels = "^0.13.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
